+++ src/main/kotlin/kuda/runtimeapi/prop/FlushGPUDirectRDMAWritesTarget.kt
@@ -1,4 +1,4 @@
-package kuda.prop
+package kuda.runtimeapi.prop
 
 enum class FlushGPUDirectRDMAWritesTarget {
     CurrentDevice
